reorganizing commandline interface to resemble mhctools and other hammerlab python tools
