+++ setup.py
@@ -1,5 +1,3 @@
-# Copyright (c) 2014. Mount Sinai School of Medicine
-#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -76,7 +74,9 @@ if __name__ == '__main__':
         ],
         long_description=readme,
         packages=find_packages(exclude="test"),
-        scripts=[
-            'scripts/topiary'
-        ],
+        entry_points={
+            'console_scripts': [
+                'topiary = topiary.cli.script:main'
+            ]
+        }
     )
