+++ topiary/converters.py
@@ -1,5 +1,3 @@
-# Copyright (c) 2015. Mount Sinai School of Medicine
-#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -17,6 +15,7 @@ Helpers to convert between data representations
 """
 
 from __future__ import print_function, division, absolute_import
+
 from collections import defaultdict
 
 import pandas as pd
