+++ topiary/lazy_ligandome_dict.py
@@ -1,5 +1,3 @@
-# Copyright (c) 2015-2017. Mount Sinai School of Medicine
-#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -11,6 +9,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from __future__ import print_function, division, absolute_import
 
 from os import listdir
 from os.path import join
