+++ topiary/rna/common.py
@@ -1,5 +1,3 @@
-# Copyright (c) 2015. Mount Sinai School of Medicine
-#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -12,8 +10,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import re
+from __future__ import print_function, division, absolute_import
 
+import re
 
 def infer_delimiter(filename, comment_char="#", n_lines=3):
     """
